move splash screen to BaseActivity, splash screen gets updated on every surface change until assets were extracted
