+++ app/src/org/koreader/launcher/JNILuaInterface.kt
@@ -39,7 +39,6 @@ internal interface JNILuaInterface {
     fun setClipboardText(text: String)
     fun setScreenBrightness(brightness: Int)
     fun setScreenOffTimeout(timeout: Int)
-    fun setShowSplashScreen(show: Boolean)
     fun setWifiEnabled(enabled: Boolean)
     fun showToast(message: String)
     fun showToast(message: String, longTimeout: Boolean)
